* main: (62 commits)
  Issue #5960: Upgrade dependencies
  Issue #5960: Upgrade dependencies
  Issue #5960: Upgrade dependencies
  Issue #5960: Upgrade dependencies
  #6076 - Display token and sentence count in the open document dialog
  #6076 - Display token and sentence count in the open document dialog
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release inception-41.0-beta-2
  [maven-release-plugin] prepare for next development iteration
  [maven-release-plugin] prepare release inception-40.5
  #5960 - Upgrade dependencies
  #5960 - Upgrade dependencies
  #5960 - Upgrade dependencies
  #6071 - Drop LIF module
  #6069 - Improve SPARQL queries against RDF4J
  #6069 - Improve SPARQL queries against RDF4J
  #6069 - Improve SPARQL queries against RDF4J
  #6069 - Improve SPARQL queries against RDF4J
  #6069 - Improve SPARQL queries against RDF4J
  #6069 - Improve SPARQL queries against RDF4J
  ...

+++ inception/inception-html-recogito-editor/src/main/ts/build.mjs
@@ -66,7 +66,9 @@ const defaults = {
 }
 
 fs.mkdirsSync(`${outbase}`)
-fs.emptyDirSync(outbase)
+if (!argv.live) {
+  fs.emptyDirSync(outbase)
+}
 
 if (argv.live) {
   const context = await esbuild.context({
